[vscode-cdm-extension] Refactor bootstrap command
